+++ src/main/java/com/github/bgomar/bgconsolelogger/chapters/Chapter.java
@@ -1,12 +1,18 @@
 package com.github.bgomar.bgconsolelogger.chapters;
 
 public class Chapter {
+    public enum Type {
+        CHAPTER, SECTION, SUBSECTION
+    }
+
     private final String title;
     private final int lineNumber;
+    private final Type type;
 
-    public Chapter(String title, int lineNumber) {
+    public Chapter(String title, int lineNumber, Type type) {
         this.title = title;
         this.lineNumber = lineNumber;
+        this.type = type;
     }
 
     public String getTitle() {
@@ -15,5 +21,9 @@ public class Chapter {
 
     public int getLineNumber() {
         return lineNumber;
+    }
+
+    public Type getType() {
+        return type;
     }
 }