+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.34] - 20-02-2025
+-  Chapter Tool Updated
+- Added option for 3 presets of chapter
+- ConsoleLogger can use empty variables again.
+- Line numbers update in file when adding a logger
+
 ## [0.0.33]
 -   🐛 Bugfix for Saving 😁👌<br><br>
     Sorry it took so long, it was quite a hard task.<br>
@@ -15,30 +21,40 @@
 - Added the Re-check line numbers function <br>
 - Fixed issue with the removal of ConsoleLogger logs <br>
 
-## [0.0.31]
+## [0.0.31] - 23-01-2025
 - tried fixing depricated issues
 
-## [0.0.30] 
+## [0.0.30] - 16-01-2025
 - Updated the logline number with a button
 
-## [0.0.29] 
+## [0.0.29] - 29-09-2024
 - changed some of the input field sizes
 - added option for second preset of loggers
 - Finally got save to work as supposed!!!
 - Fixed Issue with Code Insertion Position and Log Line Number Updates in Plugin
 - Added the Re-check line numbers function
 
-## [0.0.28]
+## [0.0.28] - 06-09-2024
 - Saving seems to work better now
 - Updated dependencies
 - removed depricated label
 - Added a cancel button
 
-## [0.0.27] 
+## [0.0.27] - 19-03-2024
 - Tried fixing the Saving issues on some IDE
 - Removed the unused settings from Tools
 
-## [0.0.21] 
+## [0.0.26] - 19-03-2024
+
+## [0.0.25] - 19-03-2024
+
+## [0.0.24] - 19-03-2024
+
+## [0.0.23] - 19-03-2024
+
+## [0.0.22] - 19-03-2024
+
+## [0.0.21] - 19-03-2024
 - 💥 Big Changes to the visuals
 - 💻 Added a toolwindow ➡️ 
 - Base64 & url En/Decoder, Fake data Generator, Cron Parser, Text -> Hash
@@ -46,10 +62,11 @@
 - 🧼 Cleanup of code and bugs - 🐛
 
 
-## [0.0.20]
+## [0.0.20] - 15-11-2023
 - 🐛 Small Bugfix: FIX: Handle cases where the editor might be missing. 
 
-## [0.0.19]
+## [0.0.19] - 23-10-2023
+
 - 🐛 Bugfix for (ctrl + alt + 1/2 ) 
 - 🤏 Small changes to Loggers defaults:
 - Added Arduino C++ (ctrl + alt + 4) = Serial.print(); 
@@ -57,55 +74,70 @@
 - Added PyCharm Py  (ctrl + alt + 6) print() 😁👌
 
 ## [0.0.18]
+
 - Upgrade dependencies
 - Added ConsoleLogger 7-8-9 for .error .group & .table
 
 ## [0.0.17]
+
 - Changes in description
 
 ## [0.0.16]
+
 - Cleanup + kotlin code style = official
 
 ## [0.0.15]
+
 - Pink Logo
 
 ## [0.0.14]
+
 - Now (ctrl + alt + 4) is for create void method from selected for Clion
 
 ## [0.0.13]
+
 - Make ConsoleLogger work with IDEA IntelliJ 233.*
 
 ## [0.0.12]
+
 - Make ConsoleLogger work with IDEA IntelliJ 233.*
 
 ## [0.0.11]
+
 - Make ConsoleLogger work with IDEA IntelliJ 232.*
 
 ## [0.0.10]
+
 - Make ConsoleLogger work with IDEA IntelliJ 2023
 
 ## [0.0.9]
+
 - Removing Color from default loggers  (CTRL + ALT + 1,2 & 3)
 - Plugin works only on IntelliJ 213.5744.223 or newer versions 223.*
 
 ## [0.0.8]
+
 - Jonas's IntelliJ is now included too
 
 ## [0.0.7]
+
 - All 6 ConsoleLoggers in Settings | Code | ConsoleLogger.
 
 ## [0.0.6]
+
 - Table bug fixed
 - Replaced console.Info() for groupCollapsed()
 - Added console.groupCollapsed()   (CTRL + ALT + 5)
 - Shorten some logs
 
 ## [0.0.5]
+
 - Fixed coloring & font size for LOG (CTRL + ALT + 1)
 - Changed order logger, INFO now (CTRL + ALT + 5)
 - fixed delete all bug
 
 ## [0.0.4]
+
 - shortcut for LOG insertion   (CTRL + ALT + 1)
 - shortcut for DEBUG insertion (CTRL + ALT + 2)
 - shortcut for WARN insertion  (CTRL + ALT + 3)
@@ -115,12 +147,15 @@
 - shortcut for deletion of all (CTRL + ALT + 0)
 
 ## [0.0.3]
+
 - Console.log ctrl+alt+1 console.warn ctrl+alt+4
 
 ## [0.0.2]
+
 - SOMETHING NEW
 
 ## [0.0.1]
+
 - BREAKING CHANGE: API class names are CamelCase. This replaces the  "lower  case version" naming.
 
 ## [Unreleased]: https://github.com/bg-omar/consolelogger/compare/v0.0.32...HEAD